moved functionality from taskcontroller.php to a new service
